Arel::Nodes::All
