+++ lib/arel/extensions.rb
@@ -22,6 +22,7 @@ require 'arel/extensions/with_ordinality'
 require 'arel/extensions/table'
 require 'arel/extensions/row'
 require 'arel/extensions/ordering'
+require 'arel/extensions/all'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
