+++ lib/arel/extensions/nodes.rb
@@ -4,10 +4,6 @@
 
 module Arel
   module Nodes
-    # postgres only: https://www.postgresql.org/docs/8.1/sql-select.html
-    class All < Arel::Nodes::Unary
-    end
-
     # postgres only: https://www.postgresql.org/docs/9.1/functions-comparisons.html
     class Any < Arel::Nodes::Unary
     end
